% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentiles.R
\name{percentiles}
\alias{percentiles}
\title{Percentiles}
\usage{
percentiles(x, ord, weight, xUp, xLow, data)
}
\arguments{
\item{x}{Variable of interest for which to calculate decile statistics}

\item{ord}{Ordering variable used to rank observations.}

\item{weight}{Survey weight variable}

\item{data}{Data frame containing the survey data}

\item{breakdown}{Numeric vector defining percentile breaks. Default creates deciles}
}
\value{
a list: total micro value by percentile, relative share owned by each group,
and number of households in each percentile
}
\description{
Calculate distributional results for all percentiles. Here we focus on deciles
but the percentiles breakdown can be customised. The ordering variable can also be selected
}
\details{
Note: raking by ordering variable is computed separately for each multiple imputation.
Some households may thus be classified in different percentiles depending on
the implicate.
}
